skills/modelarts-agent/huawei-cloud-modelarts-skill/requirements.txt
modelarts-sdk>=4.0.0